Write correct SMB2 game code to gci
Also, store the "rel" string on the stack. I originally thought this
would prevent CARDOpen() from erroring as the original char pointer was
unaligned. The actual problem was writing the correct SMB2 game code to
the GCI in gcipack, but I decided to still use "rel" on the stack
as it's more portable to other games anyway.

Note that while CARDOpen() seemed successful, the game still crashes and
I need to look into why.