ISO rel loader: use mkb2.rel_sample.rel instead of adding a new file
Adding a file with a new filename to the game apparently causes the
filename to be read into memory, causing stuff to be shifted around. In
my case it lead to the screen flashing pink while trying to reset the
graphics FIFOs, maybe because the graphics FIFOs were allocated at a
different address on the arena initially, but then when I reset their
allocation address, they were suddenly overlapping with something else.