Make some progress on bigmem modification... but it doesn't work
I tried to reset the state of memory to what it would be on a 24 MiB
console, but despite the game booting successfully, Dolphin complains
about accessing memory outside the 24 MiB range again. So it appears
that the memory locations which contain the memory size do seem to
dictate which addresses are valid and which aren't.

Unfortunately, resetting these memory addresses just before I access
extended memory did not seem to work either.

Will need to try a new approach. A couple potential avenues:
* Copy IPL memory and restore DVD BI2 location only
* Modify Dolphin to allow memory accesses outside of bounds dictated by
special memory addresses