Save entire ball struct instead of sub-region
Saving the entire ball struct seems to fix a reproducability issue on
curve bridge frames when resuming on the 06 frame from a stand-still.
Saving the whole struct isn't much bigger than what I was doing earlier
anyway; it's probably a microoptimization to reduce the amount of data
saved from the ball struct.