Refactor input system
I recently figured out most of the game's controller input data
structures, and the input system has been refactored to more effectively
leverage it. The main purpose, however, is to facilitate "exclusive
mode" which will allow the mod to receive inputs while the game
receives none. This will be used in a future menu system.