Always transition through main menu to story mode
This fixes issues with music fading and heap clearing. I think the basic gist is the game always implicitly assumes if you're loading story mode that you loaded it through the main menu. It takes a bit longer to load this way, but I think it's good enough for now (till I try to implement IWs in the menu instead)